HistoryAPI/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.100.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
async-fastapi-jwt-auth==0.6.1
gunicorn==20.1.0
redis==5.0.1
orjson==3.9.10
typer==0.9.0
passlib==1.7.4
python-multipart==0.0.6
pytest-asyncio==0.23.4
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
python-jose==3.3.0
fastapi-pagination
beanie==1.25.0